requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
langchain==0.1.9
langchain-google-genai==0.0.11
google-generativeai==0.3.2
requests==2.31.0
python-multipart==0.0.9
twilio==8.12.0
langsmith==0.1.22
pydantic==2.6.1 